Move convert_dataframes_to_lists_with_matching_pairs to plot heatmaps
